docs classes & wrappers exposed by dependency
