+++ src/dependency/core/common/exceptions.py
@@ -1,0 +1,1 @@
+class DependencyError(Exception): ...
