+++ src/dependency/core/declaration/provider.py
@@ -7,6 +7,8 @@ from dependency.core.declaration.component import Component
 from dependency.core.declaration.dependent import Dependent
 
 class Provider(ABCProvider):
+    """Provider Base Class
+    """
     def __init__(self,
             imports: list[Component],
             dependents: list[type[Dependent]],
@@ -24,7 +26,7 @@ class Provider(ABCProvider):
         self.unresolved_dependents: dict[str, list[str]] = {}
         for dependent in dependents:
             unresolved = [
-                component.__repr__() for component in dependent.imports
+                component.__repr__() for component in dependent._imports
                 if component not in self.providers
             ]
             if len(unresolved) > 0:
@@ -44,9 +46,27 @@ def provider(
         dependents: list[type[Dependent]] = [],
         provider: type[providers.Provider] = providers.Singleton
     ) -> Callable[[type], Provider]:
+    """Decorator for Provider class
+
+    Args:
+        component (type[Component]): Component class to be used as a base class for the provider.
+        imports (list[type[Component]], optional): List of components to be imported by the provider. Defaults to [].
+        dependents (list[type[Dependent]], optional): List of dependents to be declared by the provider. Defaults to [].
+        provider (type[providers.Provider], optional): Provider class to be used. Defaults to providers.Singleton.
+
+    Raises:
+        TypeError: If the wrapped class is not a subclass of Component declared base class.
+
+    Returns:
+        Callable[[type], Provider]: Decorator function that wraps the provider class.
+    """
+    # Cast due to mypy not supporting class decorators
+    _component = cast(Component, component)
+    _imports = cast(list[Component], imports)
     def wrap(cls: type) -> Provider:
-        _component = cast(Component, component)
-        _imports = cast(list[Component], imports)
+        if not issubclass(cls, _component.base_cls):
+            raise TypeError(f"Class {cls} is not a subclass of {_component.base_cls}")
+        
         provider_wrap = Provider(
             imports=_imports,
             dependents=dependents,
